moved dataclass
